fix(printf): nano-safe %lu in the networks-found string; review polish
Stage-5a review round: STR_NETWORKS_FOUND used %zu in all 34 translation
catalogs and feeds a plain snprintf, so the nano-linked combo images drew
'zu networks found' on the Wi-Fi header — translation strings are a printf
surface -Wformat never checks (finding 1). Now %lu with a call-site cast.
Also: the nano retarget script accepts both Kconfig spellings of the nano
option, fails with a diagnosis instead of a traceback when the framework
sdkconfig is missing, and names the core-gen sentinel in its exactly-one-
libc error (5-7); the ini's nano notes no longer contradict themselves on
%f and the TRAP paragraph now separates LOG-gated sites from the live
_printf_float-dependent ones (2, 4); two 512-byte decimal slips in the
lever ledger corrected — coredump is -7,996 B and silent assertions
-11,852 B, package total -79,940 B unchanged (3).

+++ src/activities/network/WifiSelectionActivity.cpp
@@ -843,7 +843,9 @@ void WifiSelectionActivity::render(RenderLock&&) {
   // STR_NETWORKS_FOUND is ~37 bytes once the Arabic translation is substituted,
   // so 32 truncated it. See ClockSyncActivity for the same class of bug.
   char countStr[64];
-  snprintf(countStr, sizeof(countStr), tr(STR_NETWORKS_FOUND), realNetworkCount);
+  // Catalogs use %lu: %zu is not understood by newlib-nano's printf (combo
+  // envs), and translation strings bypass -Wformat entirely.
+  snprintf(countStr, sizeof(countStr), tr(STR_NETWORKS_FOUND), static_cast<unsigned long>(realNetworkCount));
   GUI.drawHeader(renderer, Rect{screen.x, screen.y + metrics.topPadding, screen.width, metrics.headerHeight},
                  tr(STR_WIFI_NETWORKS), countStr);
   GUI.drawSubHeader(
